add prototype

+++ src/main/kotlin/media_item/PodcastBuilder.kt
@@ -4,7 +4,7 @@ import media_item.implementations.Podcast
 class PodcastBuilder : MediaItemBuilder() {
     override var title: String = ""
     override var creationDate: String = ""
-    var duration: Int = 0
+    var duration: Float = 0f
     lateinit var podcaster: String
 
     override fun build(): Podcast {
